+++ HeartHub/ViewController.swift
@@ -7,8 +7,13 @@
 
 import UIKit
 
-class ViewController: UIViewController {
-
+final class ViewController: UIViewController {
+//    private let loginView = LoginView()
+//    
+//    override func loadView() {
+//        view = loginView
+//    }
+    
     
     override func viewDidLoad() {
         super.viewDidLoad()
